+++ spec/test_apps/app/config/routes.rb
@@ -3,5 +3,5 @@
 Rails.application.routes.draw do
   root 'root#index'
 
-  resource :tests, only: [:destroy]
+  resource :tests, only: %i[create destroy]
 end